fix(api): move sentry init to --import instrument file for proper auto-instrumentation
Sentry.init() must run before instrumented modules (http, prisma, pino)
are imported. In ESM, static imports are hoisted before module body code,
so calling initSentry() in server.ts ran after all dependencies loaded.

Create a dedicated instrument.ts that:
- Loads dotenv (env vars aren't available at --import time)
- Calls Sentry.init() synchronously before any app modules
- Uses require() for binary mode compatibility (graceful no-op)

Update all start/dev scripts to use --import ./dist/instrument.js.
Remove initSentry() from server.ts and all workers.
