+++ apps/api/package.json
@@ -12,18 +12,18 @@
     "pnpm": ">=9.0.0"
   },
   "scripts": {
-    "dev": "tsx watch src/index.ts",
-    "dev:alert": "tsx watch src/workers/alert-monitor.ts",
-    "dev:license": "tsx watch src/workers/license-monitor.ts",
-    "dev:update": "tsx watch src/workers/update-monitor.ts",
+    "dev": "tsx watch --import ./src/instrument.ts src/index.ts",
+    "dev:alert": "tsx watch --import ./src/instrument.ts src/workers/alert-monitor.ts",
+    "dev:license": "tsx watch --import ./src/instrument.ts src/workers/license-monitor.ts",
+    "dev:update": "tsx watch --import ./src/instrument.ts src/workers/update-monitor.ts",
     "build": "prisma generate && tsc --project tsconfig.build.json && tsc-alias -f",
     "build:community": "DEPLOYMENT_MODE=community prisma generate && tsc --project tsconfig.build.json && tsc-alias -f",
     "build:enterprise": "DEPLOYMENT_MODE=enterprise prisma generate && tsc --project tsconfig.build.json && tsc-alias -f && tsx scripts/obfuscate-license.ts",
     "build:production": "pnpm run db:generate && pnpm run build && tsx scripts/obfuscate-license.ts",
-    "start": "node dist/index.js",
-    "start:alert": "node dist/workers/alert-monitor.js",
-    "start:license": "node dist/workers/license-monitor.js",
-    "start:update": "node dist/workers/update-monitor.js",
+    "start": "node --import ./dist/instrument.js dist/index.js",
+    "start:alert": "node --import ./dist/instrument.js dist/workers/alert-monitor.js",
+    "start:license": "node --import ./dist/instrument.js dist/workers/license-monitor.js",
+    "start:update": "node --import ./dist/instrument.js dist/workers/update-monitor.js",
     "type-check": "tsc --noEmit",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
