+++ apps/api/src/services/sentry/index.ts
@@ -95,127 +95,6 @@ export function trackMetricDistribution(
   }
 }
 
-export function initSentry() {
-  // Only initialize Sentry if enabled and DSN is provided
-  if (!sentryConfig.enabled) {
-    logger.info("Sentry is disabled - skipping initialization");
-    return;
-  }
-
-  if (!sentryConfig.dsn) {
-    logger.warn("Sentry DSN not provided - monitoring disabled");
-    return;
-  }
-
-  // Validate DSN format - must be a valid URL and not just "https://"
-  const dsn = sentryConfig.dsn.trim();
-  if (
-    dsn === "" ||
-    dsn === "https://" ||
-    !dsn.startsWith("https://") ||
-    dsn.length < 20
-  ) {
-    logger.warn("Invalid Sentry DSN format - monitoring disabled");
-    return;
-  }
-
-  const Sentry = getSentry();
-  if (!Sentry) {
-    logger.warn("Sentry SDK not available - monitoring disabled");
-    return;
-  }
-
-  let profilingIntegration: ReturnType<
-    typeof import("@sentry/profiling-node").nodeProfilingIntegration
-  > | null = null;
-  try {
-    // eslint-disable-next-line @typescript-eslint/no-require-imports
-    const { nodeProfilingIntegration } = require("@sentry/profiling-node");
-    profilingIntegration = nodeProfilingIntegration();
-  } catch {
-    logger.info(
-      "Sentry profiling not available - skipping profiling integration"
-    );
-  }
-
-  const integrations = [
-    // HTTP integration for API monitoring
-    Sentry.httpIntegration(),
-
-    // Prisma integration for database monitoring
-    Sentry.prismaIntegration(),
-
-    // send console.log, console.warn, and console.error calls as logs to Sentry
-    Sentry.consoleLoggingIntegration({ levels: ["log", "warn", "error"] }),
-
-    // Pino integration for structured logging
-    Sentry.pinoIntegration(),
-  ];
-
-  if (profilingIntegration) {
-    integrations.unshift(profilingIntegration);
-  }
-
-  Sentry.init({
-    dsn: dsn,
-    environment: sentryConfig.environment,
-    // Enable logs to be sent to Sentry
-    enableLogs: true,
-
-    // Enable metrics to be sent to Sentry
-    enableMetrics: true,
-
-    // Performance Monitoring
-    tracesSampleRate: sentryConfig.tracesSampleRate,
-
-    // Profiling
-    profilesSampleRate: sentryConfig.profilesSampleRate,
-
-    integrations,
-
-    // Configure what gets sent to Sentry
-    beforeSend(event) {
-      // Filter out sensitive information
-      if (event.request?.headers) {
-        delete event.request.headers.authorization;
-        delete event.request.headers.cookie;
-      }
-
-      // Don't send health check errors
-      if (event.exception?.values?.[0]?.value?.includes("health")) {
-        return null;
-      }
-
-      return event;
-    },
-
-    // Filter logs before sending to Sentry
-    beforeSendLog(log) {
-      // Skip health check logs to reduce noise
-      if (
-        log.message &&
-        typeof log.message === "string" &&
-        log.message.toLowerCase().includes("health")
-      ) {
-        return null;
-      }
-
-      return log;
-    },
-
-    // Set release information
-    release: sentryConfig.release,
-
-    // Configure tags
-    initialScope: {
-      tags: {
-        component: "backend",
-        service: "qarote-api",
-      },
-    },
-  });
-}
-
 export function setSentryUser(user: SentryUser) {
   const Sentry = getSentry();
   if (!Sentry) return;
