+++ apps/api/src/workers/update-monitor.ts
@@ -1,17 +1,9 @@
 import { logger } from "@/core/logger";
 import { prisma } from "@/core/prisma";
 
-import { initSentry } from "@/services/sentry";
-
-import { sentryConfig } from "@/config";
 import { isCloudMode } from "@/config/deployment";
 
 import { updateCheckerCronService } from "@/cron/update-checker.cron";
-
-// Initialize Sentry only if enabled
-if (isCloudMode() || sentryConfig.enabled) {
-  initSentry();
-}
 
 /**
  * Update Monitor Worker Process
